lambdas/requirements.txt
# Production dependencies for Lambda functions
boto3>=1.28.0
botocore>=1.31.0
pyyaml>=6.0
jsonschema>=4.17.0
